reorganize and rename settings
